fix: address second round of Copilot review comments
- Move GitHub release fetching/downloading into internal/updater package,
  keeping internal/cli/ free of HTTP logic per AGENTS.md separation rules
- Fix --check --json dual-output: add SilentExitError so run.go exits 1
  without emitting a second JSON error envelope after the data is written
- install.sh: replace sudo sh -c interpolation with explicit sudo mkdir/mv
  to prevent command injection from user-supplied --install-dir values
- install.sh: use jq when available for robust JSON parsing; fall back to
  tightly anchored grep/sed with head -1 to avoid multi-match issues
- install.sh: extract only the rc binary from the archive instead of
  unpacking everything, preventing path traversal from untrusted archives
