+++ internal/cli/update.go
@@ -1,40 +1,18 @@
 package cli
 
 import (
-	"archive/tar"
-	"archive/zip"
-	"compress/gzip"
-	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"os"
 	"path/filepath"
 	"runtime"
-	"strings"
 	"time"
 
 	"github.com/spf13/cobra"
 
 	"github.com/revenuecat/cli/internal/tui"
+	"github.com/revenuecat/cli/internal/updater"
 )
-
-const githubReleasesURL = "https://api.github.com/repos/RevenueCat/revenuecat-cli/releases/latest"
-
-var errUpdateAvailable = errors.New("update available")
-
-type githubRelease struct {
-	TagName string        `json:"tag_name"`
-	Assets  []githubAsset `json:"assets"`
-	HTMLURL string        `json:"html_url"`
-}
-
-type githubAsset struct {
-	Name               string `json:"name"`
-	BrowserDownloadURL string `json:"browser_download_url"`
-}
 
 func newUpdateCmd() *cobra.Command {
 	var checkOnly bool
@@ -64,18 +42,17 @@ https://github.com/RevenueCat/revenuecat-cli/releases`,
 				return nil
 			}
 
-			rt.Out.Info("Checking for updates…")
 			hc := &http.Client{Timeout: 30 * time.Second}
-			release, err := fetchLatestRelease(cmd.Context(), hc)
+
+			rt.Out.Info("Checking for updates…")
+			release, err := updater.FetchRelease(cmd.Context(), hc)
 			if err != nil {
 				return fmt.Errorf("checking for updates: %w", err)
 			}
 
-			latestTag := release.TagName
-			latestVersion := strings.TrimPrefix(latestTag, "v")
+			latestVersion := release.Version()
 
-			// Guard against downgrades (e.g. snapshot/pre-release tags).
-			if !isNewer(latestVersion, currentVersion) {
+			if !updater.IsNewer(latestVersion, currentVersion) {
 				if rt.Globals.JSON {
 					return rt.Out.Render(map[string]any{
 						"current_version": currentVersion,
@@ -91,6 +68,8 @@ https://github.com/RevenueCat/revenuecat-cli/releases`,
 
 			if checkOnly {
 				if rt.Globals.JSON {
+					// Write the JSON result to stdout, then return SilentExitError so
+					// run.go exits 1 without also emitting a JSON error envelope.
 					if err := rt.Out.Render(map[string]any{
 						"current_version": currentVersion,
 						"latest_version":  latestVersion,
@@ -98,8 +77,9 @@ https://github.com/RevenueCat/revenuecat-cli/releases`,
 					}); err != nil {
 						return err
 					}
+					return &SilentExitError{Code: 1}
 				}
-				return fmt.Errorf("%w: %s (current: %s)", errUpdateAvailable, latestVersion, currentVersion)
+				return fmt.Errorf("update available: %s (current: %s)", latestVersion, currentVersion)
 			}
 
 			if !rt.Globals.AssumeYes {
@@ -113,21 +93,8 @@ https://github.com/RevenueCat/revenuecat-cli/releases`,
 				}
 			}
 
-			assetName := assetFilename(latestVersion)
-			downloadURL := ""
-			for _, a := range release.Assets {
-				if a.Name == assetName {
-					downloadURL = a.BrowserDownloadURL
-					break
-				}
-			}
-			if downloadURL == "" {
-				return fmt.Errorf("no release asset found for %s/%s (looked for %q) — download manually: %s",
-					runtime.GOOS, runtime.GOARCH, assetName, release.HTMLURL)
-			}
-
-			rt.Out.Info(fmt.Sprintf("Downloading %s…", assetName))
-			tmpPath, err := downloadAsset(cmd.Context(), hc, downloadURL, assetName)
+			rt.Out.Info(fmt.Sprintf("Downloading %s…", release.AssetName(runtime.GOOS, runtime.GOARCH)))
+			tmpPath, err := updater.DownloadBinary(cmd.Context(), hc, release, runtime.GOOS, runtime.GOARCH)
 			if err != nil {
 				return fmt.Errorf("downloading update: %w", err)
 			}
@@ -142,7 +109,7 @@ https://github.com/RevenueCat/revenuecat-cli/releases`,
 				return fmt.Errorf("resolving binary path: %w", err)
 			}
 
-			if err := installBinary(tmpPath, execPath); err != nil {
+			if err := updater.Install(tmpPath, execPath); err != nil {
 				return fmt.Errorf("installing update: %w", err)
 			}
 
@@ -160,246 +127,4 @@ https://github.com/RevenueCat/revenuecat-cli/releases`,
 
 	cmd.Flags().BoolVar(&checkOnly, "check", false, "check for updates without installing; exit 1 if update is available")
 	return cmd
-}
-
-// isNewer returns true only when latest is strictly greater than current using
-// simple semver comparison. Returns false for equal or unparseable versions so
-// we never downgrade.
-func isNewer(latest, current string) bool {
-	lp := parseSemver(latest)
-	cp := parseSemver(current)
-	if lp == nil || cp == nil {
-		// Can't parse — treat as equal to avoid unintended updates.
-		return false
-	}
-	if lp[0] != cp[0] {
-		return lp[0] > cp[0]
-	}
-	if lp[1] != cp[1] {
-		return lp[1] > cp[1]
-	}
-	return lp[2] > cp[2]
-}
-
-// parseSemver parses "X.Y.Z" or "vX.Y.Z" into [major, minor, patch].
-func parseSemver(v string) []int {
-	v = strings.TrimPrefix(v, "v")
-	parts := strings.SplitN(v, ".", 3)
-	if len(parts) != 3 {
-		return nil
-	}
-	nums := make([]int, 3)
-	for i, p := range parts {
-		// Stop at any pre-release suffix (e.g. "0-next").
-		p, _, _ = strings.Cut(p, "-")
-		n := 0
-		for _, c := range p {
-			if c < '0' || c > '9' {
-				return nil
-			}
-			n = n*10 + int(c-'0')
-		}
-		nums[i] = n
-	}
-	return nums
-}
-
-func fetchLatestRelease(ctx context.Context, hc *http.Client) (*githubRelease, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, githubReleasesURL, nil)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Accept", "application/vnd.github+json")
-	req.Header.Set("X-GitHub-Api-Version", "2022-11-28")
-	req.Header.Set("User-Agent", "revenuecat-cli")
-
-	resp, err := hc.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		body, _ := io.ReadAll(io.LimitReader(resp.Body, 512))
-		return nil, fmt.Errorf("GitHub API returned %s: %s", resp.Status, strings.TrimSpace(string(body)))
-	}
-
-	var release githubRelease
-	if err := json.NewDecoder(resp.Body).Decode(&release); err != nil {
-		return nil, err
-	}
-	return &release, nil
-}
-
-func assetFilename(version string) string {
-	goos := runtime.GOOS
-	goarch := runtime.GOARCH
-	if goos == "windows" {
-		return fmt.Sprintf("rc_%s_%s_%s.zip", version, goos, goarch)
-	}
-	return fmt.Sprintf("rc_%s_%s_%s.tar.gz", version, goos, goarch)
-}
-
-// downloadAsset fetches the asset and extracts the rc binary to a temp file.
-func downloadAsset(ctx context.Context, hc *http.Client, url, assetName string) (string, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
-	if err != nil {
-		return "", err
-	}
-	resp, err := hc.Do(req)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("download returned %s", resp.Status)
-	}
-
-	tmp, err := os.CreateTemp("", "rc-update-*")
-	if err != nil {
-		return "", err
-	}
-	tmpName := tmp.Name()
-
-	if strings.HasSuffix(assetName, ".zip") {
-		// Write zip to disk first since zip.NewReader needs io.ReaderAt.
-		_, copyErr := io.Copy(tmp, resp.Body)
-		tmp.Close()
-		if copyErr != nil {
-			os.Remove(tmpName)
-			return "", copyErr
-		}
-		if err := extractFromZip(tmpName); err != nil {
-			os.Remove(tmpName)
-			return "", err
-		}
-		return tmpName, nil
-	}
-
-	// tar.gz — extract directly into tmp
-	gz, err := gzip.NewReader(resp.Body)
-	if err != nil {
-		tmp.Close()
-		os.Remove(tmpName)
-		return "", err
-	}
-	defer gz.Close()
-
-	tr := tar.NewReader(gz)
-	for {
-		hdr, err := tr.Next()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			tmp.Close()
-			os.Remove(tmpName)
-			return "", err
-		}
-		if filepath.Base(hdr.Name) == "rc" && hdr.Typeflag == tar.TypeReg {
-			if _, err := io.Copy(tmp, tr); err != nil {
-				tmp.Close()
-				os.Remove(tmpName)
-				return "", err
-			}
-			tmp.Close()
-			return tmpName, nil
-		}
-	}
-
-	tmp.Close()
-	os.Remove(tmpName)
-	return "", fmt.Errorf("rc binary not found in archive")
-}
-
-// extractFromZip reads a zip already written to path and overwrites it with
-// the extracted rc binary.
-func extractFromZip(path string) error {
-	r, err := zip.OpenReader(path)
-	if err != nil {
-		return err
-	}
-
-	var found bool
-	var tmpName string
-	for _, f := range r.File {
-		if filepath.Base(f.Name) == "rc.exe" || filepath.Base(f.Name) == "rc" {
-			rc, err := f.Open()
-			if err != nil {
-				r.Close()
-				return err
-			}
-			tmp, err := os.CreateTemp(filepath.Dir(path), "rc-zip-*")
-			if err != nil {
-				rc.Close()
-				r.Close()
-				return err
-			}
-			_, copyErr := io.Copy(tmp, rc)
-			rc.Close()
-			tmp.Close()
-			if copyErr != nil {
-				os.Remove(tmp.Name())
-				r.Close()
-				return copyErr
-			}
-			tmpName = tmp.Name()
-			found = true
-			break
-		}
-	}
-	// Close the zip reader before renaming so we don't hold the file open.
-	r.Close()
-
-	if !found {
-		return fmt.Errorf("rc binary not found in zip archive")
-	}
-	if err := os.Rename(tmpName, path); err != nil {
-		os.Remove(tmpName)
-		return err
-	}
-	return nil
-}
-
-// installBinary atomically replaces the current binary with the extracted one.
-func installBinary(srcPath, destPath string) error {
-	if err := os.Chmod(srcPath, 0755); err != nil {
-		return err
-	}
-
-	// Write to a sibling temp file so Rename is atomic on the same filesystem.
-	dir := filepath.Dir(destPath)
-	tmp, err := os.CreateTemp(dir, ".rc-update-*")
-	if err != nil {
-		return fmt.Errorf("cannot write to %s (check permissions): %w", dir, err)
-	}
-	tmp.Close()
-	os.Remove(tmp.Name())
-
-	if err := copyFile(srcPath, tmp.Name()); err != nil {
-		return err
-	}
-	if err := os.Chmod(tmp.Name(), 0755); err != nil {
-		os.Remove(tmp.Name())
-		return err
-	}
-	return os.Rename(tmp.Name(), destPath)
-}
-
-func copyFile(src, dst string) error {
-	in, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer in.Close()
-
-	out, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-	defer out.Close()
-
-	_, err = io.Copy(out, in)
-	return err
 }
